requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
django-taggit
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
psycopg2-binary
